table in a div in listings component
